encode: make it compatible with .NET, fix #3
This makes encoding follow .NET JSON/UTF8 libraries as much as possible,
tested with

using System;
using System.IO;
using System.Text.Json;

public class Program
{
        public static void Main()
        {
                string controls = "\x00\x01";
                MemoryStream ms = new MemoryStream();
                Utf8JsonWriter writer = new Utf8JsonWriter(ms, new JsonWriterOptions
            {
                SkipValidation = true,
            });
                writer.WriteStringValue(controls);
                writer.Flush();
                byte[] res = ms.ToArray();
                string result = System.Text.Encoding.UTF8.GetString(res);
                Console.WriteLine(result);
        }
}

which is similar to what Neo does. Can't comment on differences much because
that's just the way .NET is, it doesn't use \ufffd, it escapes some controls in
a different way, it just does what it wants and we have to follow here to be
compatible.

Notice that C# `\x` escaping differs from Go significantly if you want to play
with it (it's `\xH[H][H][H]` instead of `\xHH`). It also interprets bytes a bit
differently.

Signed-off-by: Roman Khimov <roman@nspcc.ru>